Rename .club package to .memedit
