+++ imgui-binding/src/main/java/imgui/extension/memedit/MemoryEditorSizes.java
@@ -1,4 +1,4 @@
-package imgui.extension.club;
+package imgui.extension.memedit;
 
 import imgui.binding.ImGuiStructDestroyable;
